+++ src/utils/colors.tsx
@@ -5,17 +5,25 @@ const _MULT_1 = 1.30
 const _MULT_2 = 1.15
 const _MULT_3 = 1.00
 
-export const getColor1 = (color: string): string => {
-	return shadeColor(color, _MULT_1)
+// On this get colors, the first one is the background color, and the next one is the contrast text color
+export const getColor1 = (color: string): [string, string] => {
+	const backgroundColor = shadeColor(color, _MULT_1)
+	const textColor = isOkWithWhite(hexRgb(backgroundColor)) ? 'white' : 'black'
+	return [backgroundColor, textColor]
 }
 
-export const getColor2 = (color: string): string => {
-	return shadeColor(color, _MULT_2)
+export const getColor2 = (color: string): [string, string] => {
+	const backgroundColor = shadeColor(color, _MULT_2)
+	const textColor = isOkWithWhite(hexRgb(backgroundColor)) ? 'white' : 'black'
+	return [backgroundColor, textColor]
 }
 
-export const getColor3 = (color: string): string => {
-	return shadeColor(color, _MULT_3)
+export const getColor3 = (color: string): [string, string] => {
+	const backgroundColor = shadeColor(color, _MULT_3)
+	const textColor = isOkWithWhite(hexRgb(backgroundColor)) ? 'white' : 'black'
+	return [backgroundColor, textColor]
 }
+// --
 
 const shadeColor = (color: string, amount: number): string => {
 	if (color === '') {